+++ src/index.js
@@ -1,5 +1,6 @@
 import "./components.css";
 export * from "./react/YoupuUI.jsx";
 export * from "./react/AppDialog.jsx";
+export { MultiSelect } from "./react/MultiSelect.jsx";
 export { default as TrendAreaChart } from "./react/TrendAreaChart.jsx";
 export { default as ContributionDonutChart } from "./react/ContributionDonutChart.jsx";